Revert "Merge branch 'master' into xbee-lte"
This reverts commit 56b84ea60c22d759c78634f0d99e32cfd933429a, reversing
changes made to cd072c898ca0c62ffc012a3db25fdcf6a769d5b3.
