+++ Libraries/FlexCAN_OldVersion/FlexCAN.h
@@ -1,8 +1,6 @@
 // -------------------------------------------------------------
-// a simple Arduino Teensy 3.1/3.2/3.5/3.6 CAN driver
+// a simple Arduino Teensy3.1 CAN driver
 // by teachop
-// dual CAN support for MK66FX1M0 and updates for MK64FX512 by Pawelsky
-// RTR related code by H4nky84
 //
 #ifndef __FLEXCAN_H__
 #define __FLEXCAN_H__
@@ -12,7 +10,6 @@
 typedef struct CAN_message_t {
   uint32_t id; // can identifier
   uint8_t ext; // identifier is extended
-  uint8_t rtr; // remote transmission request packet type
   uint8_t len; // length of data
   uint16_t timeout; // milliseconds, zero will disable waiting
   uint8_t buf[8];
@@ -29,10 +26,9 @@ class FlexCAN
 {
 private:
   struct CAN_filter_t defaultMask;
-  uint32_t flexcanBase;
 
 public:
-  FlexCAN(uint32_t baud = 125000, uint8_t id = 0, uint8_t txAlt = 0, uint8_t rxAlt = 0);
+  FlexCAN(uint32_t baud = 125000);
   void begin(const CAN_filter_t &mask);
   inline void begin()
   {
