+++ Playground/send_serial_can/send_serial_can.ino
@@ -1,4 +1,4 @@
-#include <HyTech_FlexCAN.h>
+#include <FlexCAN.h>
 #include "HyTech17.h"
 
 #define XBEE_PKT_LEN 15
