+++ Utilities/can_send_and_receive_test_FlexCAN35/can_send_and_receive_test_FlexCAN35.ino
@@ -33,6 +33,8 @@ void loop() {
             Serial.print(" ");
         }
         Serial.println();
+        digitalWrite(13, HIGH);
+        timer_light.reset();
     }
     while (Can0.read(msg)) { // Receive a message on CAN
         Serial.print("Received 0x");
