+++ Vehicle/Charge_Control_Unit_FH/Charge_Control_Unit_FH.ino
@@ -2,7 +2,7 @@
  * HyTech 2018 Formula Hybrid Vehicle Charge Control Unit
  */
 
-#include <HyTech_FlexCAN.h>
+#include <FlexCAN.h>
 #include "HyTech_CAN.h"
 #include <Metro.h>
