+++ Vehicle/Charge_Control_Unit_FSAE/Charge_Control_Unit_FSAE.ino
@@ -2,7 +2,7 @@
  * HyTech 2018 FSAE Vehicle Charge Control Unit
  */
 
-#include <HyTech_FlexCAN.h>
+#include <FlexCAN.h>
 #include "HyTech_CAN.h"
 #include <Metro.h>
